Merging trynow tests  (#179)
* Try Now section tests

everything has passed and even the last step has an implementation where it checks the button click and if the input has focus. If there are any issues let me know

* Try Now section tests

everything has passed and even the last step has an implementation where it checks the button click and if the input has focus. If there are any issues let me know